remove upper pin of typer
Typer was pinned with a `^0.9.0` dependency, which Poetry transforms into `(>=0.9.0,<0.10.0)`.
However, this leads to problems for projects that need later versions of Typer.
For example, diversity and vLLM will not be compatible in the same environment since because vLLM has typer>=0.12.3 as a transient dependency.
It doesn't seem like diversity will break with later versions of typer, so this commit therefore removes the upper pin.